chore: moved files to new utils directory
